Move Celeritas files into its own directory
